pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quiz_generator"
version = "0.4.0"
description = "COMP 582 Project"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "gradio==6.0.1",
    "spacy==3.7.5",
    "nltk==3.8.1",
    "scikit-learn==1.3.2",
    "gensim==4.3.2",
    "numpy==1.26.2",
    "Pillow==10.1.0",
    "pytesseract==0.3.10",
    "scipy==1.12",
    "reportlab>=4.0.0",
    "python-dotenv==1.2.1",
    "groq==0.37.1"
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

